pyproject.toml
[project]
name = "facebiometrics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "face-recognition>=1.3.0",
    "face-recognition-models",
    "matplotlib>=3.10.6",
    "mysql-connector-python>=9.4.0",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "tk>=0.1.0",
    "tkcalendar>=1.6.1",
]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models" }
